requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
crispy-bootstrap5==0.3.1
Django==3.2
django-bootstrap-v5==1.0.3
django-crispy-forms==1.11.2
django-grappelli==2.15.1
django-model-utils==4.1.1
importlib-metadata==2.1.1
Pillow==8.2.0
psycopg2-binary==2.8.6
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.1
zipp==3.4.1
